Separate ARP and ND
