+++ source/FreeRTOS_ND.c
@@ -1316,4 +1316,69 @@
         return xResult;
     }
 /*-----------------------------------------------------------*/
+
+/**
+ * @brief Check whether a packet needs ND resolution if it is on local subnet. If required send an ND Solicitation.
+ *
+ * @param[in] pxNetworkBuffer The network buffer with the packet to be checked.
+ *
+ * @return pdTRUE if the packet needs ND resolution, pdFALSE otherwise.
+ */
+BaseType_t xCheckRequiresNDResolution( const NetworkBufferDescriptor_t * pxNetworkBuffer )
+{
+    BaseType_t xNeedsNDResolution = pdFALSE;
+
+    /* MISRA Ref 11.3.1 [Misaligned access] */
+    /* More details at: https://github.com/FreeRTOS/FreeRTOS-Plus-TCP/blob/main/MISRA.md#rule-113 */
+    /* coverity[misra_c_2012_rule_11_3_violation] */
+    IPPacket_IPv6_t * pxIPPacket = ( ( IPPacket_IPv6_t * ) pxNetworkBuffer->pucEthernetBuffer );
+    IPHeader_IPv6_t * pxIPHeader = &( pxIPPacket->xIPHeader );
+    IPv6_Address_t * pxIPAddress = &( pxIPHeader->xSourceAddress );
+    uint8_t ucNextHeader = pxIPHeader->ucNextHeader;
+
+    if( ( ucNextHeader == ipPROTOCOL_TCP ) ||
+       ( ucNextHeader == ipPROTOCOL_UDP ) )
+    {
+       IPv6_Type_t eType = xIPv6_GetIPType( ( const IPv6_Address_t * ) pxIPAddress );
+       FreeRTOS_debug_printf( ( "xCheckRequiresARPResolution: %pip type %s\n",
+                                ( void * ) pxIPAddress->ucBytes,
+                                ( eType == eIPv6_Global ) ? "Global" :
+                                ( eType == eIPv6_LinkLocal ) ? "LinkLocal" :
+                                ( eType == eIPv6_Loopback ) ? "Loopback" :
+                                "other" ) );
+
+       if( eType == eIPv6_LinkLocal )
+       {
+           MACAddress_t xMACAddress;
+           NetworkEndPoint_t * pxEndPoint;
+           eARPLookupResult_t eResult;
+           char pcName[ 80 ];
+
+           ( void ) memset( &( pcName ), 0, sizeof( pcName ) );
+           eResult = eNDGetCacheEntry( pxIPAddress, &xMACAddress, &pxEndPoint );
+           FreeRTOS_printf( ( "xCheckRequiresARPResolution: eResult %s with EP %s\n", ( eResult == eARPCacheMiss ) ? "Miss" : ( eResult == eARPCacheHit ) ? "Hit" : "Error", pcEndpointName( pxEndPoint, pcName, sizeof pcName ) ) );
+
+           if( eResult == eARPCacheMiss )
+           {
+               NetworkBufferDescriptor_t * pxTempBuffer;
+               size_t uxNeededSize;
+
+               uxNeededSize = sizeof( ICMPPacket_IPv6_t );
+               pxTempBuffer = pxGetNetworkBufferWithDescriptor( BUFFER_FROM_WHERE_CALL( 199 ) uxNeededSize, 0U );
+
+               if( pxTempBuffer != NULL )
+               {
+                   pxTempBuffer->pxEndPoint = pxNetworkBuffer->pxEndPoint;
+                   pxTempBuffer->pxInterface = pxNetworkBuffer->pxInterface;
+                   vNDSendNeighbourSolicitation( pxTempBuffer, pxIPAddress );
+               }
+
+               xNeedsNDResolution = pdTRUE;
+           }
+       }
+    }
+
+    return xNeedsNDResolution;
+}
+/*-----------------------------------------------------------*/
 #endif /* ipconfigUSE_IPv6 */
