+++ test/unit-test/FreeRTOS_ARP/FreeRTOS_ARP_utest.c
@@ -1681,86 +1681,6 @@ void test_xCheckRequiresARPResolution_NotOnLocalNetwork_InvalidHeader( void )
     TEST_ASSERT_EQUAL( pdFALSE, xResult );
 }
 
-void test_xCheckRequiresARPResolution_NotOnLocalNetwork_IPv6( void )
-{
-    struct xNetworkEndPoint xEndPoint = { 0 };
-    NetworkBufferDescriptor_t xNetworkBuffer, * pxNetworkBuffer;
-    uint8_t ucEthernetBuffer[ ipconfigNETWORK_MTU ];
-    BaseType_t xResult;
-    NetworkInterface_t xInterface;
-
-    pxNetworkBuffer = &xNetworkBuffer;
-    pxNetworkBuffer->pucEthernetBuffer = ucEthernetBuffer;
-
-    IPPacket_t * pxIPPacket = ( ( IPPacket_t * ) pxNetworkBuffer->pucEthernetBuffer );
-    IPHeader_t * pxIPHeader = &( pxIPPacket->xIPHeader );
-
-    *ipLOCAL_IP_ADDRESS_POINTER = 0xABCD1234;
-
-    xEndPoint.ipv4_settings.ulNetMask = 0xFFFFFF00;
-    xNetworkBuffer.pxEndPoint = &xEndPoint;
-
-    /* Make sure there is no match. */
-    pxIPHeader->ulSourceIPAddress = ~( *ipLOCAL_IP_ADDRESS_POINTER & xEndPoint.ipv4_settings.ulNetMask );
-
-    uxIPHeaderSizePacket_IgnoreAndReturn( ipSIZE_OF_IPv6_HEADER );
-
-    xResult = xCheckRequiresARPResolution( pxNetworkBuffer );
-
-    TEST_ASSERT_EQUAL( pdFALSE, xResult );
-    /* =================================================== */
-
-    IPPacket_IPv6_t * pxIPPacket_V6 = ( ( IPPacket_IPv6_t * ) pxNetworkBuffer->pucEthernetBuffer );
-    IPHeader_IPv6_t * pxIPHeader_V6 = &( pxIPPacket_V6->xIPHeader );
-    IPv6_Address_t * pxIPAddress = &( pxIPHeader_V6->xSourceAddress );
-    pxIPHeader_V6->ucNextHeader = ipPROTOCOL_TCP;
-
-    xIPv6_GetIPType_ExpectAnyArgsAndReturn( eIPv6_LinkLocal );
-    eNDGetCacheEntry_ExpectAnyArgsAndReturn( eARPCacheHit );
-
-    xResult = xCheckRequiresARPResolution( pxNetworkBuffer );
-
-    TEST_ASSERT_EQUAL( pdFALSE, xResult );
-    /* =================================================== */
-
-    pxIPHeader_V6->ucNextHeader = ipPROTOCOL_UDP;
-
-    xIPv6_GetIPType_ExpectAnyArgsAndReturn( eIPv6_LinkLocal );
-    eNDGetCacheEntry_ExpectAnyArgsAndReturn( eARPCacheHit );
-
-    xResult = xCheckRequiresARPResolution( pxNetworkBuffer );
-
-    TEST_ASSERT_EQUAL( pdFALSE, xResult );
-    /* =================================================== */
-
-    xIPv6_GetIPType_ExpectAnyArgsAndReturn( eIPv6_LinkLocal );
-    eNDGetCacheEntry_ExpectAnyArgsAndReturn( eARPCacheMiss );
-    NetworkBufferDescriptor_t xTempBuffer = { 0 };
-    pxGetNetworkBufferWithDescriptor_ExpectAnyArgsAndReturn( &xTempBuffer );
-    vNDSendNeighbourSolicitation_Expect( &xTempBuffer, pxIPAddress );
-
-    xResult = xCheckRequiresARPResolution( pxNetworkBuffer );
-
-    TEST_ASSERT_EQUAL( pdTRUE, xResult );
-    /* =================================================== */
-
-    xIPv6_GetIPType_ExpectAnyArgsAndReturn( eIPv6_LinkLocal );
-    eNDGetCacheEntry_ExpectAnyArgsAndReturn( eARPCacheMiss );
-    pxGetNetworkBufferWithDescriptor_ExpectAnyArgsAndReturn( NULL );
-
-    xResult = xCheckRequiresARPResolution( pxNetworkBuffer );
-
-    TEST_ASSERT_EQUAL( pdTRUE, xResult );
-    /* =================================================== */
-
-    xIPv6_GetIPType_ExpectAnyArgsAndReturn( eIPv6_SiteLocal );
-
-    xResult = xCheckRequiresARPResolution( pxNetworkBuffer );
-
-    TEST_ASSERT_EQUAL( pdFALSE, xResult );
-    /* =================================================== */
-}
-
 void test_xCheckRequiresARPResolution_OnLocalNetwork_NotInCache( void )
 {
     struct xNetworkEndPoint xEndPoint = { 0 };
