+++ test/unit-test/FreeRTOS_Routing/FreeRTOS_Routing_stubs.c
@@ -42,6 +42,27 @@ uint32_t ulStubFreeRTOS_inet_ntop_TargetSize;
 const char * pcStubFreeRTOS_inet_ntop_TargetCopySource;
 uint32_t ulStubFreeRTOS_inet_ntop_CopySize;
 
+/** @brief The expected IP version and header length coded into the IP header itself. */
+#define ipIP_VERSION_AND_HEADER_LENGTH_BYTE    ( ( uint8_t ) 0x45 )
+
+UDPPacketHeader_t xDefaultPartUDPPacketHeader =
+{
+    /* .ucBytes : */
+    {
+        0x11, 0x22, 0x33, 0x44, 0x55, 0x66,  /* Ethernet source MAC address. */
+        0x08, 0x00,                          /* Ethernet frame type. */
+        ipIP_VERSION_AND_HEADER_LENGTH_BYTE, /* ucVersionHeaderLength. */
+        0x00,                                /* ucDifferentiatedServicesCode. */
+        0x00, 0x00,                          /* usLength. */
+        0x00, 0x00,                          /* usIdentification. */
+        0x00, 0x00,                          /* usFragmentOffset. */
+        ipconfigUDP_TIME_TO_LIVE,            /* ucTimeToLive */
+        ipPROTOCOL_UDP,                      /* ucProtocol. */
+        0x00, 0x00,                          /* usHeaderChecksum. */
+        0x00, 0x00, 0x00, 0x00               /* Source IP address. */
+    }
+};
+
 /* ======================== Stub Callback Functions ========================= */
 
 const char * pcStubFreeRTOS_inet_ntop( BaseType_t xAddressFamily,
